# performance: use instanceof instead of isFail() / isSuccess()