@staabm 2021-03-20
I feel we could get the most out of it, if we could either reduce `take1` to an absolute minimum regarding overhead, or alternatively try to achieve parsing by doing multiple chars at a time instead of 1 by 1 (not sure this is possible).

my assumptions stemp just from profilling source code. take it with a grain of salt, as I have zero experience building parser and haven't used parsica at all yet.